requirements.txt
duckduckgo-search>=6.2.10
numpy>2.0.0
pandas>=2.2.2
playwright>=1.47.0
beautifulsoup4>=4.12.3
requests>=2.31.0
requests-html>=0.10.0
openai>=1.51.0
crawl4ai>=0.3.14
selenium>=4.25.0
tldextract>=5.1.2
undetected-chromedriver>=3.5.5
httpx>=0.27.2
PyYAML>=6.0.2
tqdm>=4.66.5
chromadb>=0.4.24
tiktoken>=0.7.0
streamlit>=1.39.0
fastapi>=0.115.0
uvicorn>=0.30.6
fastmcp>=0.9.0
plotly>=5.24.1
google-generativeai>=0.5.0
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
dnspython>=2.4.0
googlesearch-python>=1.2.3